requirements.txt
ipykernel
ipywidgets
python-dotenv
langchain
langchain-community
groq
langchain-groq
google-generativeai
langchain-google-genai
google-genai
mistralai
langchain-mistralai
cohere
langchain-cohere
langchain_experimental
duckduckgo_search
wikipedia
ddgs
pandas-datareader
langchain_openai
langgraph
graphviz
scikit-learn
matplotlib
sentence-transformers
pinecone
langchainhub
langchain_huggingface
langchain_chroma
langchain_pinecone
pypdf
